+++ src/services/tampermonkey.ts
@@ -1,31 +1,8 @@
-export const GM_xmlhttpRequest: (request: {
-  method?: "GET" | "POST";
-  url: string;
-  headers?: Record<string, unknown>;
-  responseType?: XMLHttpRequest["responseType"];
-  data?: BodyInit | null;
-  onprogress?: (event: { loaded: number; total?: number }) => void;
-  onload?: (response: GmXhr) => void;
-  onerror?: (response: GmXhr) => void;
-  onabort?: (response: GmXhr) => void;
-}) => XMLHttpRequest = module.config().GM_xmlhttpRequest;
+// @deno-types="tampermonkey"
+import type {} from "tampermonkey";
 
-type GmXhr =
-  & {
-    responseHeaders: Record<string, string>;
-  }
-  & Pick<
-    XMLHttpRequest,
-    | "readyState"
-    | "status"
-    | "statusText"
-    | "response"
-    | "responseXML"
-    | "responseText"
-  >;
-
-declare const module: {
-  config: (() => {
-    GM_xmlhttpRequest: (() => XMLHttpRequest);
-  });
+export const setGmXhr = (xmlhttpRequest: typeof gmXhr) => {
+  gmXhr = xmlhttpRequest;
 };
+
+export let gmXhr: typeof GM_xmlhttpRequest | undefined;