+++ src/vendors/stitches.ts
@@ -1,3 +1,3 @@
-import { createStitches } from "@stitches/react";
+import { createStitches } from "../deps.ts";
 
 export const { styled, css, keyframes } = createStitches({});