+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# 8.0.0
+
+- Upgrade dependencies. The following is example of resolved packages.
+
+```
+// @resource       @stitches/react  https://cdn.jsdelivr.net/npm/@stitches/react@1.2.8/dist/index.cjs
+// @resource       fflate           https://cdn.jsdelivr.net/npm/fflate@0.7.3/lib/browser.cjs
+// @resource       object-assign    https://cdn.jsdelivr.net/npm/object-assign@4.1.1/index.js
+// @resource       react            https://cdn.jsdelivr.net/npm/react@18.2.0/cjs/react.production.min.js
+// @resource       react-dom        https://cdn.jsdelivr.net/npm/react-dom@18.2.0/cjs/react-dom.production.min.js
+// @resource       scheduler        https://cdn.jsdelivr.net/npm/scheduler@0.23.0/cjs/scheduler.production.min.js
+```
+
 ### 7.0.2
 
 - Upgrade dependencies